+++ src/system/expander/expander.js
@@ -1,0 +1,14 @@
+class Expander {
+  constructor(transformer, environment) {
+    this.transformer = transformer;
+    this.environment = environment;
+  }
+
+  to_write() {
+    return "#<Expander>"
+  }
+}
+
+const isExpander = obj => obj instanceof Expander;
+
+export { Expander, isExpander };
