+++ src/system/expander/syntactic_closure.js
@@ -1,0 +1,16 @@
+class SyntacticClosure {
+  // freeNamesはArrayか？
+  constructor(env, freeNames, form) {
+    this.environment = env;
+    this.freeNames = freeNames;
+    this.form = form;
+  }
+
+  to_write() {
+    return "#<SyntacticClosure>"
+  }
+}
+
+const isSyntacticClosure = obj => obj instanceof SyntacticClosure;
+
+export { SyntacticClosure, isSyntacticClosure };
